refactor(engine): Rename logging to logger
