+++ tracecat/api/routers/udfs.py
@@ -9,7 +9,7 @@ from tracecat import config, validation
 from tracecat.auth.credentials import authenticate_user_for_workspace
 from tracecat.db.engine import get_async_session
 from tracecat.db.schemas import UDFSpec
-from tracecat.logging import logger
+from tracecat.logger import logger
 from tracecat.types.api import UDFArgsValidationResponse
 from tracecat.types.auth import Role
