+++ tracecat/cases/service.py
@@ -12,7 +12,7 @@ from tracecat.db.engine import get_async_session_context_manager
 from tracecat.db.schemas import Case, CaseEvent
 from tracecat.identifiers import CaseID
 from tracecat.identifiers.workflow import WorkflowID
-from tracecat.logging import logger
+from tracecat.logger import logger
 from tracecat.types.auth import Role
