--- tracecat/logging/__init__.py
@@ -1,4 +1,0 @@
-from tracecat.logging._default import file_logger, standard_logger
-from tracecat.logging._logger import logger
-
-__all__ = ["logger", "standard_logger", "file_logger"]
