+++ tracecat/workflow/management/definitions.py
@@ -13,7 +13,7 @@ from tracecat.contexts import ctx_role
 from tracecat.db.engine import get_async_session_context_manager
 from tracecat.db.schemas import Workflow, WorkflowDefinition
 from tracecat.dsl.common import DSLInput
-from tracecat.logging import logger
+from tracecat.logger import logger
 from tracecat.types.auth import Role
 from tracecat.types.exceptions import TracecatException
 from tracecat.workflow.management.models import GetWorkflowDefinitionActivityInputs
