+++ tracecat/workspaces/service.py
@@ -15,7 +15,7 @@ from tracecat.contexts import ctx_role
 from tracecat.db.engine import get_async_session_context_manager
 from tracecat.db.schemas import Membership, Ownership, User, Workspace
 from tracecat.identifiers import OwnerID, UserID, WorkspaceID
-from tracecat.logging import logger
+from tracecat.logger import logger
 from tracecat.types.auth import AccessLevel, Role
 from tracecat.types.exceptions import TracecatException, TracecatManagementError
 from tracecat.workspaces.models import SearchWorkspacesParams, UpdateWorkspaceParams
